backend/requirements.txt
fastapi
uvicorn[standard]
pandas
scikit-learn
xgboost
joblib
scipy
python-multipart
python-jose[cryptography]
passlib[bcrypt]
python-dotenv
pytest
httpx
psycopg2-binary
sqlalchemy
slowapi
